Correccion de los controllers - "terminados"

+++ backend/app/Http/Controllers/PedidoCotizacionController.php
@@ -27,7 +27,8 @@ class PedidoCotizacionController extends Controller
      */
     public function store(StorePedidoCotizacionRequest $request, Obra $obra)
     {
-        $data = $request->all();
+        $data = $request->validated();
+
         if ($request->hasFile('archivo')) {
             $data['path'] = $request->file('archivo')
                 ->store('cotizaciones', 'public');
@@ -41,56 +42,71 @@ class PedidoCotizacionController extends Controller
         ], 201);
     }
 
+
     /**
      * Display the specified resource.
      */
     public function show(Obra $obra, PedidoCotizacion $pedido)
     {
-        
-        $p = PedidoCotizacion::with('obra')->find($pedido->id);       
+        if ($pedido->obra_id !== $obra->id) {
+            return response()->json([
+                'message' => 'Este pedido no pertenece a la obra',
+                'status' => 403
+            ], 403);
+        }
 
         return response()->json([
-            'pedido' => $p,
+            'pedido' => $pedido->load(['grupos','obra']),
             'status' => 200
-        ], 200);
+        ]);
     }
 
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(StorePedidoCotizacionRequest $request, Obra $obra, PedidoCotizacion $pedido)
-    {
-        $data = $request->all();
+    public function update(
+    StorePedidoCotizacionRequest $request,
+    Obra $obra,
+    PedidoCotizacion $pedido
+    ) {
+        if ($pedido->obra_id !== $obra->id) {
+            return response()->json([
+                'message' => 'Este pedido no pertenece a la obra',
+                'status' => 403
+            ], 403);
+        }
+
+        $data = $request->validated();
 
         if ($request->hasFile('archivo')) {
             if ($pedido->path) {
                 Storage::disk('public')->delete($pedido->path);
             }
-        
 
-        $data['path'] = $request->file('archivo')->store('cotizaciones', 'public');
+            $data['path'] = $request->file('archivo')
+                ->store('cotizaciones', 'public');
+        }
+
+        $pedido->update($data);
+
+        return response()->json([
+            'pedido' => $pedido->load(['grupos','obra']),
+            'status' => 200
+        ]);
     }
-
-    unset($data['archivo']);
-
-    $pedido->update($data);
-
-    return response()->json([
-        'pedido' => $pedido->load(['grupos','obra']),
-        'status' => 200
-    ], 200);
-}
-
 
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Obra $obra, PedidoCotizacion $pedido)
+   public function destroy(Obra $obra, PedidoCotizacion $pedido)
     {
         if ($pedido->obra_id !== $obra->id) {
-            abort(404);
+            return response()->json([
+                'message' => 'Este pedido no pertenece a la obra',
+                'status' => 403
+            ], 403);
         }
 
         if ($pedido->path) {
@@ -102,31 +118,8 @@ class PedidoCotizacionController extends Controller
         return response()->json([
             'message' => 'Pedido eliminado',
             'status' => 200
-        ], 200);
-    }
-
-    public function asignarGrupos(Request $request, PedidoCotizacion $pedido)
-    {
-        $validated = $request->validate([
-            'grupos' => 'required|array',
-            'grupos.*' => 'exists:grupos,id',
         ]);
-
-        $pedido->grupos()->syncWithoutDetaching($validated['grupos']);
-
-        return response()->json([
-            'message' => 'Grupos asignados correctamente'
-        ], 200);
     }
-    public function quitarGrupo(
-        PedidoCotizacion $pedido,
-        Grupo $grupo
-    ) {
-        $pedido->grupos()->detach($grupo->id);
 
-        return response()->json([
-            'message' => 'Grupo eliminado del pedido'
-        ], 200);
-    }
 
 }
